+++ src/app/demo/component/auth/accessdenied/accessdenied.component.ts
@@ -1,0 +1,9 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-accessdenied',
+  templateUrl: './accessdenied.component.html'
+})
+export class AccessdeniedComponent {
+
+}