+++ src/app/demo/component/auth/accessdenied/accessdenied.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { AccessdeniedRoutingModule } from './accessdenied-routing.module';
+import { AccessdeniedComponent } from './accessdenied.component';
+
+
+@NgModule({
+    imports: [
+        CommonModule,
+        AccessdeniedRoutingModule,
+
+    ],
+    declarations: [AccessdeniedComponent]
+})
+export class AccessdeniedModule { }