+++ src/app/demo/component/auth/auth-routing.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+
+@NgModule({
+    imports: [RouterModule.forChild([
+        { path: 'error', loadChildren: () => import('./error/error.module').then(m => m.ErrorModule) },
+        { path: 'access', loadChildren: () => import('./accessdenied/accessdenied.module').then(m => m.AccessdeniedModule) },
+        { path: 'login', loadChildren: () => import('./login/login.module').then(m => m.LoginModule) },
+
+    ])],
+    exports: [RouterModule]
+})
+export class AuthRoutingModule { }