+++ src/app/demo/component/auth/login/login.component.ts
@@ -1,0 +1,25 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'app-login',
+    templateUrl: './login.component.html',
+    styles:[`
+        i {
+            opacity: 0.6;
+            transition-duration: .12s;
+            color: #2196F3;
+            
+            &:hover {
+                opacity: 1;
+            }
+        }
+    `]
+})
+export class LoginComponent implements OnInit {
+
+    constructor() { }
+
+    ngOnInit(): void {
+    }
+
+}