+++ src/app/demo/component/auth/login/login.module.ts
@@ -1,0 +1,21 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import { LoginRoutingModule } from './login-routing.module';
+import { LoginComponent } from './login.component';
+import { InputTextModule } from 'primeng/inputtext';
+import { PasswordModule } from 'primeng/password';
+import { ButtonModule } from 'primeng/button';
+
+
+@NgModule({
+    imports: [
+        CommonModule,
+        LoginRoutingModule,
+        ButtonModule,
+        InputTextModule,
+
+    ],
+    declarations: [LoginComponent]
+})
+export class LoginModule { }