+++ src/app/demo/component/pages/empty/emptydemo.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'app-empty',
+    templateUrl: './emptydemo.component.html'
+})
+export class EmptyDemoComponent {}