+++ src/app/demo/component/pages/invoice/invoice.component.ts
@@ -1,0 +1,15 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'app-invoice',
+    templateUrl: './invoice.component.html'
+})
+export class InvoiceComponent {
+
+    constructor() { }
+
+    print() {
+        window.print();
+    }
+
+}